+++ bin/collator/src/command.rs
@@ -3,8 +3,9 @@ use crate::{
     cli::{Cli, RelayChainCli, Subcommand},
     local::{self, development_config},
     parachain::{
-        self, shibuya, shiden, start_shibuya_node, start_shiden_node, ShibuyaChainSpec,
-        ShidenChainSpec,
+        self, chain_spec, chain_spec::shibuya::ShibuyaChainSpec,
+        chain_spec::shiden::ShidenChainSpec, shibuya, shiden, start_shibuya_node,
+        start_shiden_node,
     },
     primitives::Block,
 };
@@ -64,8 +65,8 @@ fn load_spec(
         "" | "shiden" => Box::new(ShidenChainSpec::from_json_bytes(
             &include_bytes!("../res/shiden.raw.json")[..],
         )?),
-        "shibuya-testnet" | "testnet" => Box::new(parachain::get_shibuya_chain_spec(para_id)),
-        "shiden-testnet" => Box::new(parachain::get_shiden_chain_spec(para_id)),
+        "shibuya-testnet" | "testnet" => Box::new(chain_spec::shibuya::get_chain_spec(para_id)),
+        "shiden-testnet" => Box::new(chain_spec::shiden::get_chain_spec(para_id)),
         "shibuya" => Box::new(ShidenChainSpec::from_json_bytes(
             &include_bytes!("../res/shibuya.raw.json")[..],
         )?),
