--- bin/collator/src/parachain/chain_spec.rs
@@ -1,258 +1,0 @@
-//! Chain specifications.
-
-use cumulus_primitives_core::ParaId;
-use sc_chain_spec::ChainSpecExtension;
-use sc_service::ChainType;
-use serde::{Deserialize, Serialize};
-use shibuya_runtime::{
-    wasm_binary_unwrap, AccountId, AuraConfig, AuraId, Balance, BalancesConfig,
-    CollatorSelectionConfig, EVMConfig, GenesisConfig, ParachainInfoConfig, SessionConfig,
-    SessionKeys, ShibuyaNetworkPrecompiles, Signature, SudoConfig, SystemConfig, VestingConfig,
-    SDN,
-};
-use sp_core::{sr25519, Pair, Public};
-
-use sp_runtime::traits::{IdentifyAccount, Verify};
-
-type AccountPublic = <Signature as Verify>::Signer;
-
-/// Node `ChainSpec` extensions.
-///
-/// Additional parameters for some Substrate core modules,
-/// customizable from the chain spec.
-#[derive(Default, Clone, Serialize, Deserialize, ChainSpecExtension)]
-#[serde(rename_all = "camelCase")]
-pub struct Extensions {
-    /// The relay chain of the Parachain.
-    pub relay_chain: String,
-    /// The id of the Parachain.
-    pub para_id: u32,
-}
-
-impl Extensions {
-    /// Try to get the extension from the given `ChainSpec`.
-    pub fn try_get(chain_spec: &dyn sc_service::ChainSpec) -> Option<&Self> {
-        sc_chain_spec::get_extension(chain_spec.extensions())
-    }
-}
-
-/// Specialized `ChainSpec` for Shiden Network.
-pub type ShidenChainSpec = sc_service::GenericChainSpec<shiden_runtime::GenesisConfig, Extensions>;
-
-/// Specialized `ChainSpec` for Shibuya testnet.
-pub type ShibuyaChainSpec = sc_service::GenericChainSpec<GenesisConfig, Extensions>;
-
-/// Helper function to generate a crypto pair from seed
-fn get_from_seed<TPublic: Public>(seed: &str) -> <TPublic::Pair as Pair>::Public {
-    TPublic::Pair::from_string(&format!("//{}", seed), None)
-        .expect("static values are valid; qed")
-        .public()
-}
-
-/// Helper function to generate an account ID from seed
-fn get_account_id_from_seed<TPublic: Public>(seed: &str) -> AccountId
-where
-    AccountPublic: From<<TPublic::Pair as Pair>::Public>,
-{
-    AccountPublic::from(get_from_seed::<TPublic>(seed)).into_account()
-}
-
-fn shibuya_session_keys(aura: AuraId) -> SessionKeys {
-    SessionKeys { aura }
-}
-
-fn shiden_session_keys(aura: AuraId) -> shiden_runtime::SessionKeys {
-    shiden_runtime::SessionKeys { aura }
-}
-
-/// Gen Shibuya chain specification for given parachain id.
-pub fn get_shiden_chain_spec(para_id: u32) -> ShidenChainSpec {
-    // Alice as default
-    let sudo_key = get_account_id_from_seed::<sr25519::Public>("Alice");
-    let endowned = vec![
-        (
-            get_account_id_from_seed::<sr25519::Public>("Alice"),
-            1 << 70,
-        ),
-        (get_account_id_from_seed::<sr25519::Public>("Bob"), 1 << 70),
-    ];
-
-    ShidenChainSpec::from_genesis(
-        "Shiden Testnet",
-        "shiden",
-        ChainType::Live,
-        move || make_shiden_genesis(endowned.clone(), sudo_key.clone(), para_id.into()),
-        vec![],
-        None,
-        None,
-        None,
-        Extensions {
-            relay_chain: "tokyo".into(),
-            para_id,
-        },
-    )
-}
-
-/// Gen Shibuya chain specification for given parachain id.
-pub fn get_shibuya_chain_spec(para_id: u32) -> ShibuyaChainSpec {
-    // Alice as default
-    let sudo_key = get_account_id_from_seed::<sr25519::Public>("Alice");
-    let endowned = vec![
-        (
-            get_account_id_from_seed::<sr25519::Public>("Alice"),
-            1 << 70,
-        ),
-        (get_account_id_from_seed::<sr25519::Public>("Bob"), 1 << 70),
-    ];
-
-    ShibuyaChainSpec::from_genesis(
-        "Shibuya Testnet",
-        "shibuya",
-        ChainType::Live,
-        move || make_shibuya_genesis(endowned.clone(), sudo_key.clone(), para_id.into()),
-        vec![],
-        None,
-        None,
-        None,
-        Extensions {
-            relay_chain: "tokyo".into(),
-            para_id,
-        },
-    )
-}
-
-/// Helper function to create Shibuya GenesisConfig.
-fn make_shibuya_genesis(
-    balances: Vec<(AccountId, Balance)>,
-    root_key: AccountId,
-    parachain_id: ParaId,
-) -> GenesisConfig {
-    let authorities = vec![
-        (
-            get_account_id_from_seed::<sr25519::Public>("Alice"),
-            get_from_seed::<AuraId>("Alice"),
-        ),
-        (
-            get_account_id_from_seed::<sr25519::Public>("Bob"),
-            get_from_seed::<AuraId>("Bob"),
-        ),
-    ];
-
-    // This is supposed the be the simplest bytecode to revert without returning any data.
-    // We will pre-deploy it under all of our precompiles to ensure they can be called from
-    // within contracts.
-    // (PUSH1 0x00 PUSH1 0x00 REVERT)
-    let revert_bytecode = vec![0x60, 0x00, 0x60, 0x00, 0xFD];
-
-    GenesisConfig {
-        system: SystemConfig {
-            code: wasm_binary_unwrap().to_vec(),
-            changes_trie_config: Default::default(),
-        },
-        sudo: SudoConfig { key: root_key },
-        parachain_info: ParachainInfoConfig { parachain_id },
-        balances: BalancesConfig { balances },
-        vesting: VestingConfig { vesting: vec![] },
-        session: SessionConfig {
-            keys: authorities
-                .iter()
-                .map(|x| (x.0.clone(), x.0.clone(), shibuya_session_keys(x.1.clone())))
-                .collect::<Vec<_>>(),
-        },
-        aura: AuraConfig {
-            authorities: vec![],
-        },
-        aura_ext: Default::default(),
-        collator_selection: CollatorSelectionConfig {
-            desired_candidates: 200,
-            candidacy_bond: 32_000 * SDN,
-            invulnerables: authorities.iter().map(|x| x.0.clone()).collect::<Vec<_>>(),
-        },
-        evm: EVMConfig {
-            // We need _some_ code inserted at the precompile address so that
-            // the evm will actually call the address.
-            accounts: ShibuyaNetworkPrecompiles::<()>::used_addresses()
-                .map(|addr| {
-                    (
-                        addr,
-                        pallet_evm::GenesisAccount {
-                            nonce: Default::default(),
-                            balance: Default::default(),
-                            storage: Default::default(),
-                            code: revert_bytecode.clone(),
-                        },
-                    )
-                })
-                .collect(),
-        },
-        ethereum: Default::default(),
-    }
-}
-
-/// Helper function to create GenesisConfig.
-fn make_shiden_genesis(
-    balances: Vec<(AccountId, Balance)>,
-    root_key: AccountId,
-    parachain_id: ParaId,
-) -> shiden_runtime::GenesisConfig {
-    let authorities = vec![
-        (
-            get_account_id_from_seed::<sr25519::Public>("Alice"),
-            get_from_seed::<AuraId>("Alice"),
-        ),
-        (
-            get_account_id_from_seed::<sr25519::Public>("Bob"),
-            get_from_seed::<AuraId>("Bob"),
-        ),
-    ];
-
-    // This is supposed the be the simplest bytecode to revert without returning any data.
-    // We will pre-deploy it under all of our precompiles to ensure they can be called from
-    // within contracts.
-    // (PUSH1 0x00 PUSH1 0x00 REVERT)
-    let revert_bytecode = vec![0x60, 0x00, 0x60, 0x00, 0xFD];
-
-    shiden_runtime::GenesisConfig {
-        system: SystemConfig {
-            code: wasm_binary_unwrap().to_vec(),
-            changes_trie_config: Default::default(),
-        },
-        sudo: shiden_runtime::SudoConfig { key: root_key },
-        parachain_info: ParachainInfoConfig { parachain_id },
-        balances: shiden_runtime::BalancesConfig { balances },
-        vesting: shiden_runtime::VestingConfig { vesting: vec![] },
-        session: shiden_runtime::SessionConfig {
-            keys: authorities
-                .iter()
-                .map(|x| (x.0.clone(), x.0.clone(), shiden_session_keys(x.1.clone())))
-                .collect::<Vec<_>>(),
-        },
-        aura: shiden_runtime::AuraConfig {
-            authorities: vec![],
-        },
-        aura_ext: Default::default(),
-        collator_selection: shiden_runtime::CollatorSelectionConfig {
-            desired_candidates: 200,
-            candidacy_bond: 32_000 * SDN,
-            invulnerables: authorities.iter().map(|x| x.0.clone()).collect::<Vec<_>>(),
-        },
-        evm: EVMConfig {
-            // We need _some_ code inserted at the precompile address so that
-            // the evm will actually call the address.
-            accounts: ShibuyaNetworkPrecompiles::<()>::used_addresses()
-                .map(|addr| {
-                    (
-                        addr,
-                        pallet_evm::GenesisAccount {
-                            nonce: Default::default(),
-                            balance: Default::default(),
-                            storage: Default::default(),
-                            code: revert_bytecode.clone(),
-                        },
-                    )
-                })
-                .collect(),
-        },
-        ethereum: Default::default(),
-    }
-}
