+++ bin/collator/src/parachain/chain_spec/shibuya.rs
@@ -1,0 +1,135 @@
+//! Shibuya chain specifications.
+
+use cumulus_primitives_core::ParaId;
+use sc_service::ChainType;
+use shibuya_runtime::{
+    wasm_binary_unwrap, AccountId, AuraConfig, AuraId, Balance, BalancesConfig,
+    CollatorSelectionConfig, EVMConfig, GenesisConfig, ParachainInfoConfig, SessionConfig,
+    SessionKeys, ShibuyaNetworkPrecompiles, Signature, SudoConfig, SystemConfig, VestingConfig,
+    SDN,
+};
+use sp_core::{sr25519, Pair, Public};
+
+use sp_runtime::traits::{IdentifyAccount, Verify};
+
+use super::Extensions;
+
+/// Specialized `ChainSpec` for Shibuya testnet.
+pub type ShibuyaChainSpec = sc_service::GenericChainSpec<GenesisConfig, Extensions>;
+
+fn session_keys(aura: AuraId) -> SessionKeys {
+    SessionKeys { aura }
+}
+
+/// Gen Shibuya chain specification for given parachain id.
+pub fn get_chain_spec(para_id: u32) -> ShibuyaChainSpec {
+    // Alice as default
+    let sudo_key = get_account_id_from_seed::<sr25519::Public>("Alice");
+    let endowned = vec![
+        (
+            get_account_id_from_seed::<sr25519::Public>("Alice"),
+            1 << 70,
+        ),
+        (get_account_id_from_seed::<sr25519::Public>("Bob"), 1 << 70),
+    ];
+
+    ShibuyaChainSpec::from_genesis(
+        "Shibuya Testnet",
+        "shibuya",
+        ChainType::Live,
+        move || make_genesis(endowned.clone(), sudo_key.clone(), para_id.into()),
+        vec![],
+        None,
+        None,
+        None,
+        Extensions {
+            relay_chain: "tokyo".into(),
+            para_id,
+        },
+    )
+}
+
+/// Helper function to create Shibuya GenesisConfig.
+fn make_genesis(
+    balances: Vec<(AccountId, Balance)>,
+    root_key: AccountId,
+    parachain_id: ParaId,
+) -> GenesisConfig {
+    let authorities = vec![
+        (
+            get_account_id_from_seed::<sr25519::Public>("Alice"),
+            get_from_seed::<AuraId>("Alice"),
+        ),
+        (
+            get_account_id_from_seed::<sr25519::Public>("Bob"),
+            get_from_seed::<AuraId>("Bob"),
+        ),
+    ];
+
+    // This is supposed the be the simplest bytecode to revert without returning any data.
+    // We will pre-deploy it under all of our precompiles to ensure they can be called from
+    // within contracts.
+    // (PUSH1 0x00 PUSH1 0x00 REVERT)
+    let revert_bytecode = vec![0x60, 0x00, 0x60, 0x00, 0xFD];
+
+    GenesisConfig {
+        system: SystemConfig {
+            code: wasm_binary_unwrap().to_vec(),
+            changes_trie_config: Default::default(),
+        },
+        sudo: SudoConfig { key: root_key },
+        parachain_info: ParachainInfoConfig { parachain_id },
+        balances: BalancesConfig { balances },
+        vesting: VestingConfig { vesting: vec![] },
+        session: SessionConfig {
+            keys: authorities
+                .iter()
+                .map(|x| (x.0.clone(), x.0.clone(), session_keys(x.1.clone())))
+                .collect::<Vec<_>>(),
+        },
+        aura: AuraConfig {
+            authorities: vec![],
+        },
+        aura_ext: Default::default(),
+        collator_selection: CollatorSelectionConfig {
+            desired_candidates: 200,
+            candidacy_bond: 32_000 * SDN,
+            invulnerables: authorities.iter().map(|x| x.0.clone()).collect::<Vec<_>>(),
+        },
+        evm: EVMConfig {
+            // We need _some_ code inserted at the precompile address so that
+            // the evm will actually call the address.
+            accounts: ShibuyaNetworkPrecompiles::<()>::used_addresses()
+                .map(|addr| {
+                    (
+                        addr,
+                        pallet_evm::GenesisAccount {
+                            nonce: Default::default(),
+                            balance: Default::default(),
+                            storage: Default::default(),
+                            code: revert_bytecode.clone(),
+                        },
+                    )
+                })
+                .collect(),
+        },
+        ethereum: Default::default(),
+    }
+}
+
+type AccountPublic = <Signature as Verify>::Signer;
+
+/// Helper function to generate a crypto pair from seed
+fn get_from_seed<TPublic: Public>(seed: &str) -> <TPublic::Pair as Pair>::Public {
+    TPublic::Pair::from_string(&format!("//{}", seed), None)
+        .expect("static values are valid; qed")
+        .public()
+}
+
+/// Helper function to generate an account ID from seed
+fn get_account_id_from_seed<TPublic: Public>(seed: &str) -> AccountId
+where
+    AccountPublic: From<<TPublic::Pair as Pair>::Public>,
+{
+    AccountPublic::from(get_from_seed::<TPublic>(seed)).into_account()
+}
